requirements.txt
fastapi
uvicorn[standard]
transformers==4.45.0
Pillow
python-multipart
qdrant-client
python-dotenv
pdf2image
einops
timm
addict
matplotlib
gunicorn
easydict
sentence-transformers
open-clip-torch
